+++ pylatex/document.py
@@ -94,8 +94,8 @@ class Document(Environment):
 
         super().generate_tex(self._select_filepath(filepath))
 
-    def generate_pdf(self, filepath=None, clean=True, compiler='pdflatex',
-                     silent=True):
+    def generate_pdf(self, filepath=None, clean=True, clean_tex=True,
+                     compiler=None, compiler_args=None, silent=True):
         """Generate a pdf file from the document.
 
         Args
@@ -104,10 +104,23 @@ class Document(Environment):
             The name of the file (without .pdf), if it is `None` the
             ``default_filepath`` attribute will be used.
         clean: bool
-            Whether non-pdf files created by ``pdflatex`` must be removed.
+            Whether non-pdf files created that are created during compilation
+            should be removed.
+        clean_tex: bool
+            Also remove the generated tex file.
+        compiler: `str` or `None`
+            The name of the LaTeX compiler to use. If it is None, PyLaTeX will
+            choose a fitting one on its own. Starting with mklatex and then
+            pdflatex.
+        compiler_args: `list` or `None`
+            Extra arguments that should be passed to the LaTeX compiler. If
+            this is None it defaults to an empty list.
         silent: bool
             Whether to hide compiler output
         """
+
+        if compiler_args is None:
+            compiler_args = []
 
         filepath = self._select_filepath(filepath)
         filepath = os.path.join('.', filepath)
@@ -123,28 +136,75 @@ class Document(Environment):
 
         self.generate_tex(basename)
 
-        command = [compiler, '--interaction', 'nonstopmode',
-                   '--jobname', basename, basename + '.tex']
-
-        try:
-            output = subprocess.check_output(command)
-        except subprocess.CalledProcessError as e:
-            print(e.output.decode())
-            raise e
+        if compiler is not None:
+            compilers = ((compiler, []),)
         else:
-            if not silent:
-                print(output.decode())
+            latexmk_args = ['--pdf']
+            if silent:
+                latexmk_args += ['--silent']
 
-        if clean:
-            for ext in ['aux', 'log', 'out', 'tex']:
+            compilers = (
+                ('latexmk', latexmk_args),
+                ('pdflatex', [])
+            )
+
+        main_arguments = ['--interaction=nonstopmode', basename + '.tex']
+
+        os_error = None
+
+        for compiler, arguments in compilers:
+            command = [compiler] + arguments + compiler_args + main_arguments
+
+            try:
+                output = subprocess.check_output(command,
+                                                 stderr=subprocess.STDOUT)
+            except (OSError, IOError) as e:
+                # Use FileNotFoundError when python 2 is dropped
+                os_error = e
+
+                if os_error.errno == errno.ENOENT:
+                    # If compiler does not exist, try next in the list
+                    continue
+                raise(e)
+            except subprocess.CalledProcessError as e:
+                # For all other errors print the output and raise the error
+                print(e.output.decode())
+                raise(e)
+            else:
+                if not silent:
+                    print(output.decode())
+
+            if clean:
                 try:
-                    os.remove(basename + '.' + ext)
+                    # Try latexmk cleaning first
+                    subprocess.check_output(['latexmk', '-c', basename],
+                                            stderr=subprocess.STDOUT)
                 except (OSError, IOError) as e:
-                    # Use FileNotFoundError when python 2 is dropped
-                    if e.errno != errno.ENOENT:
-                        raise
+                    # Otherwise just remove some file extensions.
+                    extensions = ['aux', 'log', 'out', 'fls',
+                                  'fdb_latexmk']
+
+                    for ext in extensions:
+                        try:
+                            os.remove(basename + '.' + ext)
+                        except (OSError, IOError) as e:
+                            # Use FileNotFoundError when python 2 is dropped
+                            if e.errno != errno.ENOENT:
+                                raise
+
+            if clean_tex:
+                os.remove(basename + '.tex')  # Remove generated tex file
 
             rm_temp_dir()
+
+            # Compilation has finished, so no further compilers have to be
+            # tried
+            break
+
+        else:
+            # If none of the compilers worked, raise the last error
+            raise(os_error)
+
         os.chdir(cur_dir)
 
     def _select_filepath(self, filepath):